requirements.txt
# Core
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3

# Document loading
pypdf==4.3.1
python-docx==1.1.2
unstructured==0.15.13
atlassian-python-api==3.41.16   # Confluence

# Vector store — pgvector (PostgreSQL)
psycopg2-binary==2.9.9
pgvector==0.3.5
sqlalchemy==2.0.36

# API
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Utilities
python-dotenv==1.0.1
rich==13.9.2                     # pretty CLI output
typer==0.12.5                    # CLI framework
httpx==0.27.2

# Observability
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0

# Code quality
ruff==0.7.1
